requirements.txt
# 核心依赖
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.13.0
numpy>=1.24.0
pandas>=2.0.0

# 代码解析
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-go>=0.20.0

# 命令行界面
click>=8.0.0
rich>=13.0.0
colorama>=0.4.0

# 工具库
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.64.0
python-dotenv>=0.19.0

# 开发和测试
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# 可选：GPU支持
# torch[cuda]>=2.0.0 